+++ src/main/java/com/braintreegateway/CreditCardVerificationRequest.java
@@ -5,7 +5,12 @@ public class CreditCardVerificationRequest extends Request {
     private CreditCardVerificationCreditCardRequest creditCardRequest;
     private CreditCardVerificationOptionsRequest optionsRequest;
     private ExternalVaultVerificationRequest externalVaultVerificationRequest;
+    private String intendedTransactionSource;
+    private String paymentMethodNonce;
     private RiskDataVerificationRequest riskDataVerificationRequest;
+    private String threeDSecureToken;
+    private String threeDSecureAuthenticationID;
+    private VerificationThreeDSecurePassThruRequest verificationThreeDSecurePassThruRequest;
 
     public CreditCardVerificationRequest() {
     }
@@ -30,6 +35,31 @@ public class CreditCardVerificationRequest extends Request {
         return riskDataVerificationRequest;
     }
 
+    public CreditCardVerificationRequest intendedTransactionSource(String intendedTransactionSource) {
+        this.intendedTransactionSource = intendedTransactionSource;
+        return this;
+    }
+
+    public CreditCardVerificationRequest paymentMethodNonce(String paymentMethodNonce) {
+        this.paymentMethodNonce = paymentMethodNonce;
+        return this;
+    }
+
+    public CreditCardVerificationRequest threeDSecureAuthenticationID(String threeDSecureAuthenticationID) {
+        this.threeDSecureAuthenticationID = threeDSecureAuthenticationID;
+        return this;
+    }
+
+    public CreditCardVerificationRequest threeDSecureToken(String threeDSecureToken) {
+        this.threeDSecureToken = threeDSecureToken;
+        return this;
+    }
+
+    public VerificationThreeDSecurePassThruRequest verificationThreeDSecurePassThruRequest() {
+        this.verificationThreeDSecurePassThruRequest = new VerificationThreeDSecurePassThruRequest(this);
+        return verificationThreeDSecurePassThruRequest;
+    }
+
     @Override
     public String toXML() {
         return buildRequest("verification").toXML();
@@ -40,7 +70,12 @@ public class CreditCardVerificationRequest extends Request {
             .addElement("creditCard", creditCardRequest)
             .addElement("options", optionsRequest)
             .addElement("externalVault", externalVaultVerificationRequest)
-            .addElement("riskData", riskDataVerificationRequest);
+            .addElement("riskData", riskDataVerificationRequest)
+            .addElement("intendedTransactionSource", intendedTransactionSource)
+            .addElement("paymentMethodNonce", paymentMethodNonce)
+            .addElement("threeDSecureAuthenticationID", threeDSecureAuthenticationID)
+            .addElement("threeDSecureToken", threeDSecureToken)
+            .addElement("threeDSecurePassThru", verificationThreeDSecurePassThruRequest);
         return builder;
     }
 }